requirements.txt
Flask==2.0.2
gunicorn
newspaper3k==0.2.8
googletrans==3.1.0a0